Implement sharding counter scheme for computing the total donations.
https://developers.google.com/appengine/articles/sharding_counters

This change increments the counters, and stores all the pledges with
model_version=2 for tracking. We don't actually use the counters
yet. Once this is live and no more pledges are coming in without
model_version>=2, we'll count the total donations for pledges not
included in the counters, and manually add that.